requirements.txt
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
docopt==0.6.2
idna==2.10
mysql-connector-python==8.0.21
pipenv==2018.11.26
pipreqs==0.4.10
protobuf==3.12.2
pycparser==2.20
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.9
virtualenv==16.7.7
virtualenv-clone==0.5.3
yarg==0.1.9
